# packages/contracts/spec/openapi.yaml
...
openapi: 3.1.0
info:
    title: J.A.R.V.I.S. REST API
    description: >-
        Management API mounted at `/api` (plus `GET /health`) for the jarvis-v2
        monorepo. Authentication is a device-token bearer OR a cookie session
        (`jarvis_session`), with explicit `x-csrf-token` on cookie-authenticated
        state changes; cross-cutting logic (CSRF, owner checks) lives in server
        middleware, so this document describes the wire contract, not the auth
        policy. Application errors are returned as
        `{ "error": "<message>" }`; unhandled paths, methods, and server
        failures fall through to Express's default HTML error pages (a JSON
        404/500 handler lands with the runtime-validation phase).
    version: 0.2.0
    license:
        name: MIT
        identifier: MIT
servers:
    - url: /
      description: >-
          Relative to the origin serving this API document. The default local
          development origin is `http://localhost:54321` (`PORT` overrides the
          port); in production the server may be fronted by TLS, in which case
          the `jarvis_session` cookie is renamed `__Host-jarvis_session` (see
          the `session` security scheme).
# No security by default: explicit EMPTY security array so Redocly's
# `security-defined` rule is satisfied for the public operations; every
# protected operation below overrides this at the operation level.
security: []
...
paths:
...
    /api/attachments:
        post:
            operationId: attachmentUpload
            summary: Upload an image for analysis in a chat prompt
            description: >-
                Accepts a base64-encoded image and returns an id the client
                references in a chat prompt's `attachments` list (see the
                `/ws` AsyncAPI document). The bytes are validated by
                magic-byte sniffing, never by the request's declared type.
                Stored transiently with a fixed TTL — attachment bytes are not
                readable back and are swept on expiry; prompts referencing an
                expired id get an error frame. Requires authentication: guest
                sockets cannot upload, and the per-user storage and
                vision-model budgets are keyed on the caller. Cookie callers
                must send `x-csrf-token`.
            tags: [Attachments]
            security:
                - bearerToken: []
                - session: []
            parameters:
                - $ref: "#/components/parameters/XCsrfToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AttachmentUpload"
            responses:
                "201":
                    description: Attachment stored.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AttachmentCreated"
                "401":
                    description: Missing or invalid credentials.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "403":
                    description: >-
                        Missing or invalid `x-csrf-token` for a
                        cookie-authenticated request, or the bytes are not a
                        recognized image.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "413":
                    description: >-
                        Decoded bytes exceed the per-attachment cap, or the
                        caller is over their total-attachment budget. The body
                        carries the cap so a client can retry smaller.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AttachmentTooLarge"
    /api/bootstrap:
        post:
            operationId: bootstrap
            summary: Bootstrap the first-owner account
            description: >-
                Creates the initial owner account and returns a freshly issued
                device token. Gated by the `JARVIS_BOOTSTRAP_TOKEN` env secret
                (sent as `x-bootstrap-token`), which is consumed on success so a
                leftover secret can never create a second owner. Rate-limited
                per source IP.
            tags: [Bootstrap]
            parameters:
                - $ref: "#/components/parameters/XBootstrapToken"
            requestBody:
                required: true
                description: >-
                    Owner username, password, and optional device name
                    (default `cli`).
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CredentialLogin"
            responses:
                "201":
                    description: Owner account and device token created.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AuthResult"
                "400":
                    description: Malformed or out-of-range request body.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "403":
                    description: The `x-bootstrap-token` secret did not match.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "409":
                    description: >-
                        An owner already exists, or first-owner setup is
                        disabled (no `JARVIS_BOOTSTRAP_TOKEN` configured).
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "429":
                    description: Too many bootstrap attempts; try again later.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                "500":
                    description: Unexpected server error.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
components:
    securitySchemes:
        bearerToken:
            type: http
            scheme: bearer
            description: >-
                An opaque device token issued by `POST /api/bootstrap`, `POST
                /api/auth/login`, or `POST /api/devices`. It is shown exactly
                once at issue time and cannot be recovered.
        session:
            type: apiKey
            in: cookie
            name: jarvis_session
            description: >-
                The login session cookie set by `POST /api/session`
                (`HttpOnly`/`SameSite=Strict`). Under TLS it is renamed
                `__Host-jarvis_session` and flagged `Secure`; the plain
                `jarvis_session` name is used over cleartext HTTP. Cookie-
                authenticated state-changing requests must also send
                `x-csrf-token`.
    parameters:
...
        XBootstrapToken:
            name: x-bootstrap-token
            in: header
            required: false
            description: >-
                The `JARVIS_BOOTSTRAP_TOKEN` secret that gates first-owner
                setup. Optional at the shape layer on purpose: a missing or
                mismatched secret is an authorization outcome (403, or 409
                when setup is disabled), decided by the server — not a
                request-shape violation.
            schema:
                type: string
        XCsrfToken:
            name: x-csrf-token
            in: header
            required: false
            description: >-
                Required on cookie-authenticated state-changing requests —
                echo the `csrfToken` from the session response. Bearer-token
                callers omit it.
            schema:
                type: string
    schemas:
        AttachmentUpload:
            type: object
            description: >-
                A base64-encoded image upload. Only `data` is meaningful —
                the type is decided by magic-byte sniffing of the decoded
                bytes, never by a declared mime.
            required: [data]
            properties:
                data:
                    type: string
                    description: >-
                        Base64-encoded image bytes (PNG, JPEG, GIF, or WebP).
                        `maxLength` corresponds to the server's decoded-byte
                        cap (4 MiB → 5,592,408 base64 characters); oversize
                        bodies are answered `413` by the route, not by shape
                        validation. Deliberately no `pattern`: a regex over
                        megabytes of base64 is real CPU on every request, and
                        a broken encoding surfaces as a decode failure (403),
                        not a shape violation.
                    maxLength: 5592408
        AttachmentCreated:
            type: object
            description: The id to reference in a chat prompt's `attachments` list.
            required: [attachmentId]
            properties:
                attachmentId:
                    type: string
                    description: >-
                        Opaque attachment id (a 24-character base64url string
                        in the current implementation). The protocol accepts
                        ids up to 32 characters, leaving slack for a future
                        scheme.
                    minLength: 1
                    maxLength: 32
        AttachmentTooLarge:
            type: object
            description: Rejection for a too-large upload or an exhausted budget.
            required: [error, code, maxBytes]
            properties:
                error:
                    type: string
                code:
                    type: string
                    enum:
                        - ATTACHMENT_TOO_LARGE
                    description: Machine-readable reason a client branches on.
                maxBytes:
                    type: integer
                    description: The per-attachment cap in decoded bytes.
        Error:
            type: object
            description: User-safe error envelope returned by the JSON error handler.
            required: [error]
            properties:
                error:
                    type: string
        User:
            type: object
            description: Public account row as exposed by the `/api` routes.
            required: [id, username, role, disabled, createdAt]
            properties:
                id:
                    type: integer
                    readOnly: true
                    description: Stable integer row id.
                username:
                    type: string
                    minLength: 1
                    maxLength: 64
                role:
                    enum: [owner, user]
                disabled:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
                    readOnly: true
...
        IssuedDevice:
            type: object
            description: >-
                A freshly provisioned device credential: the persisted row plus
                the one-time secret token that is never stored.
            required: [id, name, prefix, token]
            properties:
                id:
                    type: integer
                    readOnly: true
                name:
                    type: string
                    minLength: 1
                    maxLength: 64
                prefix:
                    type: string
                    readOnly: true
                token:
                    type: string
                    readOnly: true
                    description: The one-time bearer secret; shown once and unrecoverable.
        AuthResult:
            type: object
            description: Response of both device-token flows (`/api/bootstrap` and `/api/auth/login`).
            required: [user, device]
            properties:
                user:
                    $ref: "#/components/schemas/User"
                device:
                    $ref: "#/components/schemas/IssuedDevice"
...
        CredentialLogin:
            type: object
            description: >-
                Credentials for device-token flows (`POST /api/bootstrap` and
                `POST /api/auth/login`). `deviceName` defaults to `cli`.
            required: [username, password]
            properties:
                username:
                    type: string
                    minLength: 1
                    maxLength: 64
                password:
                    type: string
                    minLength: 8
                    maxLength: 1024
                deviceName:
                    type: string
                    minLength: 1
                    maxLength: 64
                    default: cli
...